remove vue-class-component from devDependence and ignoure some tests when use vue3 type
